Changed all doubles used for distances to long

+++ src/main/java/edu/csu2017fa314/T15/Model/Edge.java
@@ -26,7 +26,7 @@ public class Edge {
         return sourceID;
     }
 
-    public double getDistance(){
+    public long getDistance(){
         return distance;
     }
